Drop unneeded definitions and move webhooks methods to webhook.go
